Fulfillment working with gas optimized request

+++ contracts/src/v0.8/interfaces/VRFCoordinatorV2Interface.sol
@@ -4,10 +4,10 @@ interface VRFCoordinatorV2Interface {
 
     function requestRandomWords(
         bytes32 keyHash,  // Corresponds to a particular offchain job which uses that key for the proofs
-        uint32  subId,   // A data structure for billing
-        uint16  minimumRequestConfirmations,
-        uint32  callbackGasLimit,
-        uint16 numWords  // Desired number of random words
+        uint64  subId,   // A data structure for billing
+        uint64  minimumRequestConfirmations,
+        uint64  callbackGasLimit,
+        uint64  numWords  // Desired number of random words
     )
     external
     returns (uint256 requestId);
@@ -16,10 +16,10 @@ interface VRFCoordinatorV2Interface {
         address[] memory consumers // permitted consumers of the subscription
     )
     external
-    returns (uint32 subId);
+    returns (uint64 subId);
 
     function fundSubscription(
-        uint32 subId,
+        uint64 subId,
         uint256 amount
     )
     external;
